The initial implementation of the run-length-encoder (RLE) (reworked)
This is the initial implementation of the run-length-encoder.  Some integration work, documentation, and refactoring needs to be completed.  The package and subpackages need to be streamlined so that it is straightforward to access the public API.  The public interfaces need to placed in a common place so that it is easy to build the system from the interfaces.  The interfaces all need to be merged, there should be a small number of global interfaces used in this design (any number of local interfaces can be used).

* Update rlecore.py
* Update rlecore.py
* modified entropycoder
* updated rlecore
* added cfelton/rhea in requirements.txt
...