Added DCT data flow test
Added a DCT data flow test to emulate the possible data flow through
the system.  This commit also has changed to the interfaces, the
previous simulation `assign` functions were replaced with the `next`
attribute.  The `assign` function is now a myhdl.block to be used
in the convertible HDL blocks.