Moved the reference design tests to test_ref.
The Verilog cosimulation tests, reference design tests, were moved
to a separate directory.  This was done to organize the tests.  The
cosim (ref tests) being in a separate directory can be troublesome.
The cosim needs to know the path to the Verilog files and the VPI
module.  When run from pytest this typically will be from the
test_jpeg/tests/ directory, all the paths would be relative from
here.  Currently all the paths are fixed from the tests/ dir, this
should be enhanced in the future to determine which directory the
pytest runner exectuing from and adjust the paths accordingly.